WIP: ConfigFileLoader
